// docker/back-end/src/types/client.ts
export type ClientType = {
  clientId: number;
  fullName: string;
  birthday: string;
  cpf: string;
  cnh: string;
  gender: string;
  nacionality: string;
  phoneNumber: string;
  email: string;
  profession: string;
  salary: number;
  bank: string;
  timesPurchased: number;
  password: string;
  addressId: number;
};